test(config): move config_setup fixture to worker/conftest.py for better test isolation
Relocate the config_setup pytest fixture from lib/conftest.py and sinks/conftest.py into worker/conftest.py to centralize configuration setup and teardown for tests. This improves test isolation and reduces duplication.
