+++ osprey_worker/src/osprey/worker/lib/conftest.py
@@ -4,11 +4,8 @@ from .patcher import patch_all  # isort: skip
 # please ensure this occurs before *any* other imports !
 patch_all(patch_gevent=False, patch_ddtrace=False)
 
-from typing import Any, Generator  # noqa: E402
 
-import pytest  # noqa: E402
 from osprey.engine import conftest as rules_conftest  # noqa: E402
-from osprey.worker.lib.singletons import CONFIG  # noqa: E402
 
 from .tests import test_utils  # noqa: E402
 
@@ -21,17 +18,6 @@ udf_registry = rules_conftest.udf_registry
 
 # Rules-package fixtures used for testing validators
 from _pytest.config.argparsing import Parser
-
-
-@pytest.fixture(autouse=True)  # autouse = True means automatically use for each test
-def config_setup() -> Generator[Any, None, None]:
-    CONFIG.instance().configure_from_env()
-    # yield is used here to basically split this function into two parts:
-    # all code before `yield` is the setup code (run before each test), and
-    # all code after `yield` is the teardown code (run after each test)
-    yield  # this line is where the testing happens
-    # teardown code
-    CONFIG.instance().unconfigure_for_tests()
 
 
 def pytest_addoption(parser: Parser) -> None:
